// src/api/api.ts
export interface CreateCategoryDTO {
    'name': string;
    'type': CreateCategoryDTOTypeEnum;
    'color'?: string;
}

export const CreateCategoryDTOTypeEnum = {
    Despesa: 'DESPESA',
    Receita: 'RECEITA'
} as const;

export type CreateCategoryDTOTypeEnum = typeof CreateCategoryDTOTypeEnum[keyof typeof CreateCategoryDTOTypeEnum];
// ...
export interface UpdateCategoryDTO {
    'name': string;
    'type': UpdateCategoryDTOTypeEnum;
    'color'?: string;
}

export const UpdateCategoryDTOTypeEnum = {
    Despesa: 'DESPESA',
    Receita: 'RECEITA'
} as const;

export type UpdateCategoryDTOTypeEnum = typeof UpdateCategoryDTOTypeEnum[keyof typeof UpdateCategoryDTOTypeEnum];

export interface UpdateDateTransactionDTO {
    'date': string;
}
export interface UpdateDescriptionTransactionDTO {
    'description': string;
}
export interface UpdateEmailAndUsernameDTO {
    'username': string;
    'email': string;
}
export interface UpdateEmailDTO {
    'email': string;
}
export interface UpdateNameCategoryDTO {
    'name': string;
}
export interface UpdatePasswordDTO {
    'currentPassword': string;
    'password': string;
    'confirmPassword': string;
}
export interface UpdateTransactionDTO {
    'description': string;
    'date': string;
    'amount': number;
    'category_id': string;
}
export interface UpdateTypeCategoryDTO {
    'type': UpdateTypeCategoryDTOTypeEnum;
}

export const UpdateTypeCategoryDTOTypeEnum = {
    Despesa: 'DESPESA',
    Receita: 'RECEITA'
} as const;

export type UpdateTypeCategoryDTOTypeEnum = typeof UpdateTypeCategoryDTOTypeEnum[keyof typeof UpdateTypeCategoryDTOTypeEnum];
// ...
export class CategoriaApi extends BaseAPI {
    /**
     * 
     * @param {CreateCategoryDTO} createCategoryDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public createCategory(createCategoryDTO: CreateCategoryDTO, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).createCategory(createCategoryDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} categoryId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public deleteCategory(categoryId: string, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).deleteCategory(categoryId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getAllCategories(options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).getAllCategories(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} categoryId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public getCategoryById(categoryId: string, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).getCategoryById(categoryId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} categoryId 
     * @param {UpdateCategoryDTO} updateCategoryDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateCategory(categoryId: string, updateCategoryDTO: UpdateCategoryDTO, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).updateCategory(categoryId, updateCategoryDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} categoryId 
     * @param {UpdateNameCategoryDTO} updateNameCategoryDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateCategoryName(categoryId: string, updateNameCategoryDTO: UpdateNameCategoryDTO, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).updateCategoryName(categoryId, updateNameCategoryDTO, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} categoryId 
     * @param {UpdateTypeCategoryDTO} updateTypeCategoryDTO 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public updateCategoryType(categoryId: string, updateTypeCategoryDTO: UpdateTypeCategoryDTO, options?: RawAxiosRequestConfig) {
        return CategoriaApiFp(this.configuration).updateCategoryType(categoryId, updateTypeCategoryDTO, options).then((request) => request(this.axios, this.basePath));
    }
}